Completado el Exportador y Agregado el Informe
El TP está completo con esto
