+++ Juego.cpp
@@ -1,4 +1,4 @@
-#include "juego.h"
+#include "Juego.h"
 
 //PRIVATE
 char Juego::getFichaAleatoria(){
