+++ Jugador.cpp
@@ -1,4 +1,4 @@
-#include "jugador.h"
+#include "Jugador.h"
 
 std::string Jugador::generarColorAleatorio(){
